Remove submodules, replace them with package.json references
platform-server, like platform-gnome, and unlike platform-cloud and
platform-cmdline, is meant as a "release artifact" platform that
is actually usable, and is not where actual development of Almond
occurs. As such, we don't need submodules, and we can adopt a more
traditional approach to nodejs library dependencies, with package.json
referring to our git repository, and yarn.lock locking the version
to a specific git commit.