User: fix pbkdf with modern node
Recent versions of node require the digest explicitly, because sha1 is a bad
default.

sha1 is still secure in PBKDF form (as the security of PBKDF is based on how
expensive sha is to compute rather than true collision resistance), and we
use it everywhere else so it is ok.