Transform into a multi-user assistant
Copy over the multi-user EngineManager from almond-cloud, but
run it in the same process as the frontend (which is simpler),
and avoid shared processes (as we don't have memory issues here).

Use sqlite rather than mysql for the user database.

Voice commands are not dispatched (we need speaker recognition
for that) and the frontend is broken, but it's a start!