docker: try harder fixing snowboy
Two things can happen: in one case, snowboy builds correctly the
first time, and the second time nothing happens. In the other
case, snowboy fails to build, so we go into the fallback building
manually, but because it failed to build, we also did not install
node-pre-gyp (cause that's a dep of snowboy) and fail twice.
To avoid that, with need to install snowboy's dependencies first,
without running any scripts (because scripts will try to run "prepare"
and fail, as we're running from the tarball not a git checkout).