platform: delay initialization of PulseAudio
If we initialize PulseAudio before the engine is opened, we might
do a lot of work and sync IO, and delay the event loop enough
that PulseAudio terminates the connection. This happens especially
on Raspberry Pi, which has a slow CPU and painfully slow IO.

Instead, delay touching the audio until we really need it, when
we initialize the speech conversation.