Use Travis to build docker images
While our application to Docker Open Source is under review, builds
on DockerHub are not available. Instead, set up travis to build
our docker images.

The full build is done with "docker buildx", which in theory lets
us build images for other architectures. In practice, the normal
build (changing the base image to an arch specific image) works
for ARM but fails in "docker buildx", and I'm not sure exactly why.